production/requirements.txt
openai-agents==0.0.19
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
aiokafka>=0.10.0
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-cloud-pubsub>=2.21.0
twilio>=9.1.0
pydantic>=2.7.0
pydantic[email]>=2.7.0
pgvector>=0.2.5
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
